Move even more into mock_transport
